+++ node/src/commons/config-manager.js
